Resolves #5
Remove the `create user` call in the final task as it is not idempotent
(won't work on the second run). The `grant user` command will create user
on first run.

Used a file based semaphor to allow use of the MySQL `root` user to create
a non-standard super user on the first run only. After that, removal of the
`root` user would cause an error. Would love to know if there is a better way
to have a task only execute on the first run of a playbook.

Reduced the scope of the user `mysqluser` to be only the `electronics`
schema to reduce blast radius from that attack vector. Ideally this would be
scoped much tighter to something like "insert, update, delete, select" from
is current "ALL PRIVILEGES". I chose to make a note rather than dig through
the application to see what it would actually need since it is issuing DDL.

Removed the debian account as a gratuitous attack vector.

Change the port from 3306 to 1521 (standard oracle port) because I was asked
to and I like irony!